refactor(core): moved `proxySignal` to core/reactivity
