+++ projects/core/reactivity/debounced/index.ts
@@ -5,11 +5,12 @@ import {
   type Signal,
   type WritableSignal,
 } from '@angular/core';
-import { proxySignal, setupContext } from '@signality/core/internal';
+import { setupContext } from '@signality/core/internal';
 import { toValue } from '@signality/core/utilities';
 import type { MaybeSignal, SignalValue, WithInjector } from '@signality/core/types';
 import { debounceCallback } from '@signality/core/scheduling/debounce-callback';
 import { watcher } from '@signality/core/reactivity/watcher';
+import { proxySignal } from '@signality/core/reactivity/proxy-signal';
 
 export type DebouncedOptions<T> = CreateSignalOptions<T> & WithInjector;
