+++ projects/core/reactivity/index.ts
@@ -1,3 +1,4 @@
 export * from '@signality/core/reactivity/debounced';
+export * from '@signality/core/reactivity/proxy-signal';
 export * from '@signality/core/reactivity/throttled';
 export * from '@signality/core/reactivity/watcher';
