+++ projects/core/router/title/index.ts
@@ -3,8 +3,8 @@ import { ActivatedRoute } from '@angular/router';
 import { Title } from '@angular/platform-browser';
 import { toSignal } from '@angular/core/rxjs-interop';
 import { filter } from 'rxjs';
-import { proxySignal, setupContext } from '@signality/core/internal';
 import type { WithInjector } from '@signality/core/types';
+import { proxySignal } from '@signality/core/reactivity/proxy-signal';
 
 export type TitleOptions = CreateSignalOptions<string> & WithInjector;
